+++ packages/runtime/src/runtime/footer.tsx
@@ -1,5 +1,5 @@
 import type { EnsembleFooterModel } from "@ensembleui/react-framework";
-import { Column } from "../widgets";
+import { Column } from "../widgets/Column";
 
 interface EnsembleFooterProps {
   footer?: EnsembleFooterModel;
@@ -12,7 +12,7 @@ export const EnsembleFooter: React.FC<EnsembleFooterProps> = ({ footer }) => {
 };
 
 const prepareFooter = (
-  footer: EnsembleFooterModel | undefined
+  footer: EnsembleFooterModel | undefined,
 ): React.ReactElement | undefined => {
   if (!footer) return;
 
@@ -22,12 +22,12 @@ const prepareFooter = (
       position: "static",
       display: "flex",
       justifyContent: "center",
-      width: footer?.styles?.width || "100%",
-      backgroundColor: footer?.styles?.backgroundColor || "white",
-      height: footer?.styles?.height || "56px",
+      width: footer.styles?.width || "100%",
+      backgroundColor: footer.styles?.backgroundColor || "white",
+      height: footer.styles?.height || "56px",
       bottom: 0,
     },
   };
 
-  return <Column {...defaultStyles} children={footer.children} />;
+  return <Column {...defaultStyles}>{footer.children}</Column>;
 };