+++ packages/runtime/src/shared/types.ts
@@ -1,9 +1,11 @@
 import type { EnsembleWidget, Expression } from "@ensembleui/react-framework";
 
-export interface EnsembleWidgetProps {
+export type EnsembleWidgetStyles = Record<string, string | number>;
+
+export interface EnsembleWidgetProps<T = EnsembleWidgetStyles> {
   id?: string;
   [key: string]: unknown;
-  styles?: Record<string, string | number>;
+  styles?: T;
 }
 
 export type BaseTextProps = {
@@ -34,19 +36,6 @@ export type IconProps = {
     margin?: number | string;
   } & HasBorder;
 } & EnsembleWidgetProps;
-
-export interface GridViewStyles {
-  horizontalTileCount?: number;
-  horizontalGap?: number;
-  verticalGap?: number;
-}
-
-export type SearchStyles = {
-  width?: number;
-  height?: number;
-  margin?: number | string;
-  backgroundColor?: string;
-} & HasBorder;
 
 // composable types
 export interface HasBorder {