+++ packages/runtime/src/widgets/Charts/index.tsx
@@ -12,9 +12,9 @@ import {
   type ChartOptions,
 } from "chart.js";
 import React, { cloneElement } from "react";
-import type { Expression } from "@ensembleui/react-framework";
+import { useWidgetId, type Expression } from "@ensembleui/react-framework";
 import { WidgetRegistry } from "../../registry";
-import type { EnsembleWidgetProps } from "../../util/types";
+import type { EnsembleWidgetProps } from "../../shared/types";
 import { BarChart } from "./BarChart";
 import { DoughnutChart } from "./DoughnutChart";
 import { StackBarChart } from "./StackBarChart";
@@ -60,12 +60,17 @@ const tabsConfig = {
 
 export const Chart: React.FC<ChartProps> = (props) => {
   const { type } = props;
+  const { rootRef } = useWidgetId(props.id);
 
   if (!type) {
     return <b>Chart type missing</b>;
   }
 
-  return cloneElement(tabsConfig[type], { ...props });
+  return (
+    <div ref={rootRef} style={{ height: "100%", width: "100%" }}>
+      {cloneElement(tabsConfig[type], { ...props })}
+    </div>
+  );
 };
 
 WidgetRegistry.register("Chart", Chart);