+++ packages/runtime/src/widgets/Form/Checkbox.tsx
@@ -1,7 +1,7 @@
 import { Form as AntForm, Checkbox } from "antd";
 import { useState } from "react";
 import { useRegisterBindings } from "@ensembleui/react-framework";
-import type { EnsembleWidgetProps } from "../../util/types";
+import type { EnsembleWidgetProps } from "../../shared/types";
 import { WidgetRegistry } from "../../registry";
 import type { FormInputProps } from "./types";