+++ packages/runtime/src/widgets/Markdown.tsx
@@ -3,8 +3,8 @@ import React, { useState } from "react";
 import ReactMarkdown from "react-markdown";
 import { useRegisterBindings } from "@ensembleui/react-framework";
 import { WidgetRegistry } from "../registry";
-import type { BaseTextProps } from "../util/types";
-import { getTextAlign } from "../util/utils";
+import type { BaseTextProps } from "../shared/types";
+import { getTextAlign } from "../shared/styles";
 
 export type MarkdownProps = {
   // to be added more