+++ packages/runtime/src/widgets/Text.tsx
@@ -2,8 +2,8 @@ import { useRegisterBindings } from "@ensembleui/react-framework";
 import { useState } from "react";
 import { Typography } from "antd";
 import { WidgetRegistry } from "../registry";
-import type { BaseTextProps } from "../util/types";
-import { getTextAlign } from "../util/utils";
+import type { BaseTextProps } from "../shared/types";
+import { getTextAlign } from "../shared/styles";
 
 export interface TextStyles {
   fontSize?: string | number;